requirements.txt
streamlit
pandas
psycopg2
python-dotenv
fuzzywuzzy
sqlalchemy
python-Levenshtein
numpy